+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Create background tasks middlewares
+
 ### 0.22.1 - 2020-04-12
 
  - Fix factory when has middlewares and the controller has no inputs